chore - added checking on user on cards and offer modifications

+++ src/main/java/com/wcs/mtgbox/collection/domain/service/impl/CollectionCardServiceImpl.java
@@ -10,15 +10,13 @@ import com.wcs.mtgbox.collection.domain.entity.CardLanguage;
 import com.wcs.mtgbox.collection.domain.entity.CardQuality;
 import com.wcs.mtgbox.collection.domain.entity.UserCard;
 import com.wcs.mtgbox.collection.domain.service.CollectionCardService;
-import com.wcs.mtgbox.collection.infrastructure.exception.CardLanguageNotFoundErrorException;
-import com.wcs.mtgbox.collection.infrastructure.exception.CardNotFoundErrorException;
-import com.wcs.mtgbox.collection.infrastructure.exception.CardQualityNotFoundErrorException;
-import com.wcs.mtgbox.collection.infrastructure.exception.UserCardNotFoundErrorException;
+import com.wcs.mtgbox.collection.infrastructure.exception.*;
 import com.wcs.mtgbox.collection.infrastructure.repository.CardLanguageRepository;
 import com.wcs.mtgbox.collection.infrastructure.repository.CardQualityRepository;
 import com.wcs.mtgbox.collection.infrastructure.repository.CardRepository;
 import com.wcs.mtgbox.collection.infrastructure.repository.UserCardRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -50,7 +48,8 @@ public class CollectionCardServiceImpl implements CollectionCardService {
         this.cardMapper = cardMapper;
     }
     @Override
-    public List<CollectionCardDto> saveCollectionCards(List<AddCollectionCardDto> addCardList) {
+    @Transactional
+    public List<CollectionCardDto> saveCollectionCards(List<AddCollectionCardDto> addCardList, Long tokenUserId) {
 
         List<CollectionCardDto> newCollectionCards = new ArrayList<CollectionCardDto>();
 
@@ -63,7 +62,7 @@ public class CollectionCardServiceImpl implements CollectionCardService {
             Card cardInfo = cardRepository.findById(dbCard.get().getId())
                     .orElseThrow(CardNotFoundErrorException::new);
 
-            UserCard userCard = this.saveUserCard(cardInfo,card);
+            UserCard userCard = this.saveUserCard(cardInfo, card, tokenUserId);
 
             CollectionCardDto collectionCard = cardMapper.userCardEntityToCollectionCardDto(userCard);
 
@@ -77,7 +76,10 @@ public class CollectionCardServiceImpl implements CollectionCardService {
         return Optional.of(cardRepository.save(cardMapper.addCollectionCardDtoToCardEntity(newCard)));
     }
 
-    private UserCard saveUserCard(Card cardInfo, AddCollectionCardDto addCardDto) {
+    private UserCard saveUserCard(Card cardInfo, AddCollectionCardDto addCardDto, Long tokenUserId) {
+        if (!addCardDto.getUserInfo().getUserId().equals(tokenUserId)) {
+            throw new UserNotAuthorizedErrorException();
+        }
         User user = userRepository.findById(addCardDto.getUserInfo().getUserId())
                 .orElseThrow(UserNotFoundErrorException::new);
         CardQuality cardQuality = cardQualityRepository.findByName(addCardDto.getUserInfo().getQualityName())
@@ -106,15 +108,21 @@ public class CollectionCardServiceImpl implements CollectionCardService {
     }
 
     @Override
-    public void removeUserCardId(Long userCardId) {
+    public void removeUserCardId(Long userCardId, Long tokenUserId) {
         UserCard userCard = userCardRepository.findById(userCardId)
                 .orElseThrow(UserCardNotFoundErrorException::new);
+        if (!userCard.getUser().getId().equals(tokenUserId)) {
+            throw new UserNotAuthorizedErrorException();
+        }
         userCardRepository.delete(userCard);
     }
 
     @Override
-    public void removeUserCardsByUserCardIds(List<Long> userCardIdList) {
-        List<UserCard> userCards = userCardRepository.findAllById(userCardIdList);
+    public void removeUserCardsByUserCardIds(List<Long> userCardIdList, Long tokenUserId) {
+        List<UserCard> userCards = userCardRepository.findAllById(userCardIdList)
+                .stream()
+                .filter(userCard -> userCard.getUser().getId().equals(tokenUserId))
+                .toList();
         if (userCards.isEmpty()) {
             throw new UserCardNotFoundErrorException();
         }
